[lexical-rich-text][lexical-a11y] Feature: Announce heading changes by default in rich text
The announcer moves from @lexical/a11y into @lexical/rich-text, beside
HeadingNode, and RichTextExtension depends on it - so a rich text editor
announces heading changes without anyone adding anything, and `disabled`
switches it off. @lexical/rich-text now depends on @lexical/a11y rather
than the reverse, so an announcer never drags a node package into an
editor that did not ask for one.
