+++ packages/lexical-playground/src/App.tsx
@@ -10,7 +10,6 @@ import {
   EditorModeAnnounceExtension,
   FocusManagerExtension,
   FocusTrapExtension,
-  HeadingAnnounceExtension,
   HistoryAnnounceExtension,
   RovingTabIndexExtension,
 } from '@lexical/a11y';
@@ -266,7 +265,6 @@ const AppExtension = /* @__PURE__ */ defineExtension({
     HistoryExtension,
     HistoryAnnounceExtension,
     EditorModeAnnounceExtension,
-    HeadingAnnounceExtension,
     KeywordsExtension,
     HashtagExtension,
     DateTimeExtension,
