+++ packages/lexical-rich-text/flow/LexicalRichText.js.flow
@@ -6,6 +6,7 @@
  *
  * @flow strict
  */
+import type {NamedSignalsOutput} from '@lexical/extension';
 import type {
   DOMConversionMap,
   EditorConfig,
@@ -96,3 +97,15 @@ export type SerializedHeadingNode = {
 
 
 declare export var RichTextExtension: LexicalExtension<RichTextConfig, "@lexical/rich-text", void, void>;
+
+export type HeadingAnnounceExtensionConfig = {
+  created: string,
+  destroyed: string,
+  disabled: boolean,
+};
+declare export var HeadingAnnounceExtension: LexicalExtension<
+  HeadingAnnounceExtensionConfig,
+  '@lexical/rich-text/HeadingAnnounce',
+  NamedSignalsOutput<HeadingAnnounceExtensionConfig>,
+  void,
+>;
