+++ packages/lexical-rich-text/package.json
@@ -38,6 +38,7 @@
     }
   },
   "dependencies": {
+    "@lexical/a11y": "workspace:*",
     "@lexical/clipboard": "workspace:*",
     "@lexical/dragon": "workspace:*",
     "@lexical/extension": "workspace:*",
