+++ packages/lexical-rich-text/src/index.ts
@@ -1939,6 +1939,10 @@ export function registerRichText(
 }
 
 export {
+  HeadingAnnounceExtension,
+  type HeadingAnnounceExtensionConfig,
+} from './HeadingAnnounceExtension';
+export {
   type RichTextConfig,
   RichTextExtension,
   RichTextImportExtension,
